% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.ASD.replicats.R
\name{average.ASD.replicats}
\alias{average.ASD.replicats}
\title{Calcule la moyenne et l'écart-type des réplicats mesurés avec l'ASD}
\usage{
average.ASD.replicats(file.list, verbose = FALSE)
}
\arguments{
\item{file.list}{est la liste des noms de fichier ASD.}

\item{verbose}{est indique si on veut voir les noms des fichiers qui sont lus. Par défaut verbose=FALSE.}
}
\value{
Elle retourne une liste avec les variables suivantes.
Lum= une matrice de luminances normalisées.
L.norm= Luminance normalisée par le temps d'intégration (moyenne).
L.norm.sd= Luminance normalisée par le temps d'intégration (ecart-type).
waves= Les longueur d'onde. The wavelenght vector;
DateTime= La date et l'heure de l'acquisition en format POSIXct pour chaque acquisition.
IntTime= Le temps d'intégration en millisecondes pour chaque acquisition.
}
\description{
Calcule la moyenne et l'écart-type des réplicats mesurés avec l'ASD
}
\author{
Simon Bélanger
}
