% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.ASD.rhow.R
\name{compute.ASD.rhow}
\alias{compute.ASD.rhow}
\title{Compute the water leaving reflectance from ASD data}
\usage{
compute.ASD.rhow(raw.asd, rho.panel = 0.985, quantile.prob = 0.5,
  VERBOSE = TRUE)
}
\arguments{
\item{raw.asd}{is a long list returned by \code{\link{merge.ASD.radiances.for.rhow}} containing all
the data necessary to compute the diffuse water reflectance. That include the raw radiance data,
the sun-viewing geometry, the location, the time and the wind speed.}

\item{rho.panel}{is the reflectivity of the spectralon panel. The Default is 0.985}

\item{quantile.prob}{is a value (between 0.25 to 1) for the maximum quantile probability
for which the surface radiance (Lt) values will be discarded.
The quantile probability is the value at which the probability
of the random variable being less than or equal to this value.
For example, a value of 0.5 means that every Lt value above the
50\% quantile probability will be discarded, while a value of 0.8 means
that only the values above the 80\% quantile probability will be discarded.
The rational is to eliminate outliers resulting from sun glitters, foam, sea ice, etc.
The default value is 0.5 to eliminate Lt value above the 50\% quantile probability.}

\item{VERBOSE}{is a bolean to output the processing steps in the terminal (Default is TRUE).}
}
\description{
This function compute the remote sensing reflectance for a water surface.
}
\details{
See User's Guide (in french) for details
}
\author{
Simon Bélanger
}
