% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.NDVI.R
\name{compute.NDVI}
\alias{compute.NDVI}
\title{Calcule le NDVI à partir de la réflectance hyperspectrale pour différents
capteurs satellitaires (Landsat4or5-TM, Landsat7-ETM+, Landsat8-OLI,
Sentinel2-MSI, Pleiades and WorldView)}
\usage{
compute.NDVI(waves, R, sensor = "OLI")
}
\arguments{
\item{waves}{est un vecteur de longueur d'onde}

\item{R}{est la reflectance spectrale}

\item{sensor}{est le nom du capteur satellitaire,
i.e. TM, ETM+, OLI, MSI, PLEIADES ou WV.
(Par défaut sensor = "OLI)}
}
\value{
Retourne la valeur du NDVI
}
\description{
L'indice de végétation NDVI (Normalized Difference Vegetation Index)
est la différence entre la réflectance dans le proche infrarouge
et la réflectance dans le rouge normalisé par la somme des deux.
Le NDVI varie théoriquement entre -1 et 1.
La fonction considère les différences spectrales entre les capteurs.
}
\author{
Simon Bélanger
}
