% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.ASD.rho.R
\name{export.ASD.rho}
\alias{export.ASD.rho}
\title{Export spectral refelctance in an ASCII file}
\usage{
export.ASD.rho(asd, file.out = "asd.rho.txt")
}
\arguments{
\item{asd}{est une liste retourné par la fonction \code{\link{compute.ASD.rho}}}

\item{file.out}{est le nom du fichier ASCII ou les données seront écrites}
}
\description{
Export spectral refelctance in an ASCII file
}
\author{
Simon Bélanger
}

