% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.ASD.rhow.DB.R
\name{generate.ASD.rhow.DB}
\alias{generate.ASD.rhow.DB}
\title{Generate a data base for rhow from the ASD}
\usage{
generate.ASD.rhow.DB(path = "./", mission = "XXX")
}
\arguments{
\item{path}{is the path where the file directories.for.ASD.dat containing the data folders
to merge in the data base.}

\item{mission}{is a string for the name of the mission. It will be used for the file names of the output.}
}
\value{
It returns a list object named ASD.DB containing a matrix of rhow (rhow.m) and vectors for
StationID, data, lat, lon sunzen, rhow.Method.

The object ASD.DB is saved in RData format. The data are also saved in ASCII (.dat with comma separator)
and a figure showing the measured rho_w spectra of the data base is produced.
}
\description{
Generate a data base for rhow from the ASD
}
