% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.file.names.ASD.R
\name{generate.file.names.ASD}
\alias{generate.file.names.ASD}
\title{generate a vector of ASD file names ending by *.asd.txt}
\usage{
generate.file.names.ASD(basename, ASD_start, ASD_end, ADD_UNDERSCORE = TRUE)
}
\arguments{
\item{basename}{is a character string corresponding to the base of the file names
produced by the RS3 software  (e.g. "StationX_00010.asd.txt"; the basename is "StationX").}

\item{ASD_start}{is a number corresponding to the first file of a set of ASD measurement
(e.g. "StationX_00010.asd.txt"; the number is 10).}

\item{ASD_end}{is a number corresponding to the last file of a set of ASD measurement.}

\item{ADD_UNDERSCORE}{is a logical parameter indicating whether or not the basename is
separated from the file number in the name. By default, the basename is separated from the file
number by an underscore (ADD_UNDERSCORE=TRUE).}
}
\description{
generate a vector of ASD file names ending by *.asd.txt
}
