% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.rho550.R
\name{get.rho550}
\alias{get.rho550}
\title{Retreive sea-surface reflectance from a look-up-tables}
\usage{
get.rho550(thetaV, delta.phi, windspeed, thetaS)
}
\arguments{
\item{thetaV}{is the viewing zenith angle}

\item{delta.phi}{is the sensor azimuth angle relative to the sun}

\item{windspeed}{is the wind speed in m/s.}

\item{thetaS}{is the sun zenith angle}
}
\description{
Interpolate within a 4D look-up-table (LUT) of the sea-surface reflectance
computed by Curt Mobley (Mobley, 1999, 2015).
}
\details{
The function performs a linear interpolation in a 4D LUT.
It uses the \code{\link{approx3d}} function from the oce package.
The current LUT is decribed in details in Mobley, App. Opt (1999), which was updated by
Mobley, App. Opt (2015) to include improve sea surface wave statistic and polarisation
}
\examples{
get.rho550(thetaV=35, delta.phi=120, windspeed=5.5, thetaS=55)

}
\author{
Simon Bélanger
}
