% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.ASD.radiances.for.rhow.R
\name{merge.ASD.radiances.for.rhow}
\alias{merge.ASD.radiances.for.rhow}
\title{Merge ASD radiances files from the surface, the sky and the spectralon
panel taken to determine the diffuse above-water reflectance}
\usage{
\method{merge}{ASD.radiances.for.rhow}(Ltot, Lsky, Lpanel,
  StationID = "StationX", lat, lon, DateTime, ThetaV, Dphi, Windspeed)
}
\arguments{
\item{Ltot}{is a list returned by the function \code{\link{average.ASD.replicats}}
containing the radiances of the surface (Ltot)}

\item{Lsky}{is a list returned by the function \code{\link{average.ASD.replicats}}
containing the radiances of the sky (Lsky)}

\item{Lpanel}{is a list returned by the function \code{\link{average.ASD.replicats}}
containing the radiances of the spectralon (Lpanel)}

\item{StationID}{is the station ID}

\item{lat}{is the latitude in degree of the measurements}

\item{lon}{is the longitude in degree of the measurements}

\item{DateTime}{is the averaged time of the measurements in POSIXct format}

\item{ThetaV}{is the viewing zenith angle}

\item{Dphi}{is the sensor azimuth angle relative to the sun}

\item{Windspeed}{is the wind speed in m/s.}
}
\value{
Returns a list containing all the radiance data of the three
type of measurements required for the Rrs calculation of the water
diffuse reflectance. In addition, ancillary data list needed to compute the Rrs
is also provided (lat,lon, DateTime, ThetaS, ThetaV and DeltaPhi)
}
\description{
Merge ASD radiances files from the surface, the sky and the spectralon
panel taken to determine the diffuse above-water reflectance
}
\author{
Simon Bélanger
}
