% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ASD.rho.R
\name{plot.ASD.rho}
\alias{plot.ASD.rho}
\title{Plot ASD reflectance}
\usage{
\method{plot}{ASD.rho}(asd, type.surf = "Surface", PNG = FALSE)
}
\description{
Plot the ASD reflectance return by the function compute.ASD.rho
}
\details{
@param asd is a list return by compute.ASD.rho
}

