% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ASD.rhow.R
\name{plot.ASD.rhow}
\alias{plot.ASD.rhow}
\title{Produit des figures pour la réflectance marine}
\usage{
plot.ASD.rhow(asd, PNG = FALSE, RADIANCES = FALSE)
}
\description{
@param asd est une liste produite par la fonction \code{\link{compute.ASD.rhow}}
@param PNG est une variable booléenne (TRUE ou FALSE) qui permet de produire un fichier png.
Par défaut PNG=FALSE
@param RADIANCES est une variable booléenne (TRUE ou FALSE) qui permet de produire une figure
avec les mesures de luminances et la réflectance de la surface et du ciel. Par défaut RADIANCES=FALSE
}
\details{
@author Simon Bélanger
}
