% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possol.R
\name{possol}
\alias{possol}
\title{Compute the Sun position from the location and time on Earth}
\usage{
possol(month, jday, tu, xlon, xlat)
}
\arguments{
\item{month}{is an integer for the month number (1 to 12)}

\item{jday}{is the number of the day in the month}

\item{tu}{is the time in decimal in UTC}

\item{xlon}{is the longitude in decimal degree}

\item{xlat}{is the latitude in decimal degree}
}
\value{
It returns a vector of two numeric for the sun zenith angle and the azimuthal angle of the sun.
}
\description{
Compute the Sun position from the location and time on Earth
}
\author{
Bernard Gentilly, LOV
}
