% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.ASD.R
\name{process.ASD}
\alias{process.ASD}
\title{Process a ASD data folder}
\usage{
process.ASD(dirdat, PNG = FALSE, ADD_UNDERSCORE = TRUE)
}
\arguments{
\item{dirdat}{is the current directory path contaning the files to process.}

\item{PNG}{is a logical parameter indicating whether or not diagnostic plots are saved in PNG format.
Two types of plot are produced by the function \code{\link{plot.ASD.rhow}} and saved in a sub-folder
/PNG/ that will be created in working directory. Default is PNG=FALSE.}

\item{ADD_UNDERSCORE}{is a logical parameter indicating whether or not the basename of the ASD files to processed are
separated from the file number in the name. By default, the basename is separated from the file
number by an underscore (e.g. "StationX_00001.asd.txt"). The default is ADD_UNDERSCORE=TRUE.}
}
\value{
The function returns a list containing the rhow data.
The same list (rhow) is saved into a rhow.RData file in dirdat.
}
\description{
This function is called by the higher level function \code{\link{ASD.go}}.
It can be also called in the command line to process a given data directory.
}
\details{
The function first looks into the cast.info.dat file found
in the dirdat and check whether all the mandatory fields are present.
If not, the processing may be stop or a default value is taken.
Next the ASD files are merged using \code{\link{merge.ASD.radiances.for.rhow}}.
Finally, the function calls \code{\link{compute.ASD.rhow}} to
compute rhow spectra using various methods.
}
\seealso{
\code{\link{compute.ASD.rhow}} and \code{\link{ASD.go}}
}
\author{
Simon Bélanger
}
