% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ASD.R
\name{read.ASD}
\alias{read.ASD}
\title{Lit les données ASCII de l'ASD produites par le logiciel ViewSpec Pro.
Reads ASD ASCII files exported from ViewSpec Pro.}
\usage{
read.ASD(file)
}
\arguments{
\item{file}{est le nom complet du fichier ASD. Is the ASD file name}
}
\value{
Elle retourne une liste avec les variables suivantes.
Returns a list with the following variables:
L.norm= Luminance normalisée par le temps d'intégration. Raw radiance normalized by the integration time;
waves= Les longueur d'onde. The wavelenght vector;
DateTime= La date et l'heure de l'acquisition en format POSIXct.
The date and time in POSIXtc format;
IntTime= Le temps d'intégration en millisecondes. The integration time in milliseconds
}
\description{
Lit les données ASCII de l'ASD produites par le logiciel ViewSpec Pro.
Reads ASD ASCII files exported from ViewSpec Pro.
}
\author{
Simon Bélanger
}
